requirements.txt
pyvi
ftfy
tqdm
flask
regex
numpy
pandas
faiss-cpu
clip
rapidfuzz
matplotlib
flask-cors
opencv-python
langdetect==1.0.9
googletrans==3.1.0a0
sentence_transformers
transformers
scikit-learn==1.4.0
open_clip_torch

# Merge with sis-master
Pillow
tensorflow>=2.0.0